requirements.txt
fastapi
uvicorn[standard]
python-dotenv
groq
pinecone
supabase
pydantic
